@@ -109,12 +109,11 @@ fn test_language_corpus(
             .join("corpus");
     }
 
-    let error_corpus_file = error_corpus_dir.join(&format!("{}_errors.txt", language_name));
-    let template_corpus_file =
-        template_corpus_dir.join(&format!("{}_templates.txt", language_name));
+    let error_corpus_file = error_corpus_dir.join(format!("{language_name}_errors.txt"));
+    let template_corpus_file = template_corpus_dir.join(format!("{language_name}_templates.txt"));
     let main_tests = parse_tests(&corpus_dir).unwrap();
-    let error_tests = parse_tests(&error_corpus_file).unwrap_or(TestEntry::default());
-    let template_tests = parse_tests(&template_corpus_file).unwrap_or(TestEntry::default());
+    let error_tests = parse_tests(&error_corpus_file).unwrap_or_default();
+    let template_tests = parse_tests(&template_corpus_file).unwrap_or_default();
     let mut tests = flatten_tests(main_tests);
     tests.extend(flatten_tests(error_tests));
     tests.extend(flatten_tests(template_tests).into_iter().map(|mut t| {
@@ -122,7 +121,7 @@ fn test_language_corpus(
         t
     }));
 
-    let mut skipped = skipped.map(|x| HashMap::<&str, usize>::from_iter(x.iter().map(|x| (*x, 0))));
+    let mut skipped = skipped.map(|x| x.iter().map(|x| (*x, 0)).collect::<HashMap<&str, usize>>());
 
     let language_path = if subdir.is_empty() {
         language_name.to_string()
@@ -136,7 +135,7 @@ fn test_language_corpus(
     let dump_edits = env::var("TREE_SITTER_DUMP_EDITS").is_ok();
 
     if log_seed {
-        println!("  start seed: {}", start_seed);
+        println!("  start seed: {start_seed}");
     }
 
     println!();
@@ -168,7 +167,7 @@ fn test_language_corpus(
                 println!("Incorrect initial parse for {test_name}");
                 print_diff_key();
                 print_diff(&actual_output, &test.output);
-                println!("");
+                println!();
                 return false;
             }
 
@@ -208,7 +207,7 @@ fn test_language_corpus(
                 }
 
                 if log_seed {
-                    println!("   {test_index}.{trial:<2} seed: {}", seed);
+                    println!("   {test_index}.{trial:<2} seed: {seed}");
                 }
 
                 if dump_edits {
@@ -255,7 +254,7 @@ fn test_language_corpus(
                     println!("Incorrect parse for {test_name} - seed {seed}");
                     print_diff_key();
                     print_diff(&actual_output, &test.output);
-                    println!("");
+                    println!();
                     return false;
                 }
 
@@ -276,14 +275,15 @@ fn test_language_corpus(
         }
     }
 
-    if failure_count > 0 {
-        panic!("{} {} corpus tests failed", failure_count, language_name);
-    }
+    assert!(
+        failure_count == 0,
+        "{failure_count} {language_name} corpus tests failed"
+    );
 
     if let Some(skipped) = skipped.as_mut() {
         skipped.retain(|_, v| *v == 0);
 
-        if skipped.len() > 0 {
+        if !skipped.is_empty() {
             println!("Non matchable skip definitions:");
             for k in skipped.keys() {
                 println!("  {k}");
@@ -298,7 +298,7 @@ fn test_feature_corpus_files() {
     let test_grammars_dir = fixtures_dir().join("test_grammars");
 
     let mut failure_count = 0;
-    for entry in fs::read_dir(&test_grammars_dir).unwrap() {
+    for entry in fs::read_dir(test_grammars_dir).unwrap() {
         let entry = entry.unwrap();
         if !entry.metadata().unwrap().is_dir() {
             continue;
@@ -326,7 +326,7 @@ fn test_feature_corpus_files() {
                 continue;
             }
 
-            eprintln!("test language: {:?}", language_name);
+            eprintln!("test language: {language_name:?}");
 
             let expected_message = fs::read_to_string(&error_message_path)
                 .unwrap()
@@ -335,24 +335,17 @@ fn test_feature_corpus_files() {
                 let actual_message = e.to_string().replace("\r\n", "\n");
                 if expected_message != actual_message {
                     eprintln!(
-                        "Unexpected error message.\n\nExpected:\n\n{}\nActual:\n\n{}\n",
-                        expected_message, actual_message
+                        "Unexpected error message.\n\nExpected:\n\n{expected_message}\nActual:\n\n{actual_message}\n",
                     );
                     failure_count += 1;
                 }
             } else {
-                eprintln!(
-                    "Expected error message but got none for test grammar '{}'",
-                    language_name
-                );
+                eprintln!("Expected error message but got none for test grammar '{language_name}'",);
                 failure_count += 1;
             }
         } else {
             if let Err(e) = &generate_result {
-                eprintln!(
-                    "Unexpected error for test grammar '{}':\n{}",
-                    language_name, e
-                );
+                eprintln!("Unexpected error for test grammar '{language_name}':\n{e}",);
                 failure_count += 1;
                 continue;
             }
@@ -364,7 +357,7 @@ fn test_feature_corpus_files() {
             let tests = flatten_tests(test);
 
             if !tests.is_empty() {
-                eprintln!("test language: {:?}", language_name);
+                eprintln!("test language: {language_name:?}");
             }
 
             for test in tests {
@@ -384,7 +377,7 @@ fn test_feature_corpus_files() {
                     } else {
                         print_diff_key();
                         print_diff(&actual_output, &test.output);
-                        println!("");
+                        println!();
                         false
                     }
                 });
@@ -396,12 +389,11 @@ fn test_feature_corpus_files() {
             }
         }
     }
-    if failure_count > 0 {
-        panic!("{} corpus tests failed", failure_count);
-    }
+
+    assert!(failure_count == 0, "{failure_count} corpus tests failed");
 }
 
-fn check_consistent_sizes(tree: &Tree, input: &Vec<u8>) {
+fn check_consistent_sizes(tree: &Tree, input: &[u8]) {
     fn check(node: Node, line_offsets: &Vec<usize>) {
         let start_byte = node.start_byte();
         let end_byte = node.end_byte();
@@ -449,7 +441,7 @@ fn check_consistent_sizes(tree: &Tree, input: &Vec<u8>) {
 
     let mut line_offsets = vec![0];
     for (i, c) in input.iter().enumerate() {
-        if *c == '\n' as u8 {
+        if *c == b'\n' {
             line_offsets.push(i + 1);
         }
     }
@@ -457,7 +449,7 @@ fn check_consistent_sizes(tree: &Tree, input: &Vec<u8>) {
     check(tree.root_node(), &line_offsets);
 }
 
-fn check_changed_ranges(old_tree: &Tree, new_tree: &Tree, input: &Vec<u8>) -> Result<(), String> {
+fn check_changed_ranges(old_tree: &Tree, new_tree: &Tree, input: &[u8]) -> Result<(), String> {
     let changed_ranges = old_tree.changed_ranges(new_tree).collect::<Vec<_>>();
     let old_scope_sequence = ScopeSequence::new(old_tree);
     let new_scope_sequence = ScopeSequence::new(new_tree);
@@ -473,13 +465,12 @@ fn check_changed_ranges(old_tree: &Tree, new_tree: &Tree, input: &Vec<u8>) -> Re
     for range in &changed_ranges {
         if range.end_byte > byte_range.end || range.end_point > point_range.end {
             return Err(format!(
-                "changed range extends outside of the old and new trees {:?}",
-                range
+                "changed range extends outside of the old and new trees {range:?}",
             ));
         }
     }
 
-    old_scope_sequence.check_changes(&new_scope_sequence, &input, &changed_ranges)
+    old_scope_sequence.check_changes(&new_scope_sequence, input, &changed_ranges)
 }
 
 fn set_included_ranges(parser: &mut Parser, input: &[u8], delimiters: Option<(&str, &str)>) {
@@ -532,9 +523,9 @@ fn get_parser(session: &mut Option<util::LogSession>, log_filename: &str) -> Par
     if *LOG_ENABLED {
         parser.set_logger(Some(Box::new(|log_type, msg| {
             if log_type == LogType::Lex {
-                eprintln!("  {}", msg);
+                eprintln!("  {msg}");
             } else {
-                eprintln!("{}", msg);
+                eprintln!("{msg}");
             }
         })));
     } else if *LOG_GRAPH_ENABLED {
